Solves (for now) the problem, use canvas.Refresh instead of the diagram clear + redraw combo.
Update issue 5
The solution seems to be to use canvas.Refresh instead of the diagram clear + redraw combo.  The problem with
  self.GetDiagram().Clear(dc)
  self.GetDiagram().Redraw(dc)
is that the clear() only clears the physical visible area and the scrolled off area still has content, so that when you scroll to it you get old rubbish there. e.g. if the shape that was moved was in the scrolled off area, you will see double - the old position and the new.